Factor image handling out into a mixin
